2024-02-07 Merge

+++ Assets/02. Scripts/Data/ResourceObjectSaveData.cs
@@ -14,7 +14,7 @@ public struct ResourceObjectState
     }
 
     public int state;
-    public float remainingTime;
+    public int remainingTime;
 }
 
 [System.Serializable]
